+++ example/lib/model/user/person/person.dart
@@ -9,13 +9,9 @@ part 'person.g.dart';
 @JsonSerializable(explicitToJson: true)
 @immutable
 class Person {
-  @JsonKey(name: 'firstName', required: true, includeIfNull: false)
+  @JsonKey(name: 'firstName', required: true)
   final String firstName;
-  @JsonKey(
-      name: 'gender',
-      required: true,
-      includeIfNull: false,
-      unknownEnumValue: Gender.X)
+  @JsonKey(name: 'gender', required: true, unknownEnumValue: Gender.x)
   final Gender gender;
 
   const Person({